route conflict on get tickets for phase
