+++ db/tickets_test.go
@@ -223,111 +223,3 @@ func TestGetTicket(t *testing.T) {
 		}
 	})
 }
-
-func TestGetTicketsByPhase(t *testing.T) { // create person
-	InitTestDB()
-
-	// create person
-	now := time.Now()
-
-	person := Person{
-		Uuid:        uuid.New().String(),
-		OwnerPubKey: "testfeaturepubkey",
-		OwnerAlias:  "testfeaturealias",
-		Description: "testfeaturedescription",
-		Created:     &now,
-		Updated:     &now,
-		Deleted:     false,
-	}
-
-	// create person
-	TestDB.CreateOrEditPerson(person)
-
-	workspace := Workspace{
-		Uuid:    uuid.New().String(),
-		Name:    "Test tickets space",
-		Created: &now,
-		Updated: &now,
-	}
-
-	// create workspace
-	TestDB.CreateOrEditWorkspace(workspace)
-
-	workspaceFeatures := WorkspaceFeatures{
-		Uuid:          uuid.New().String(),
-		WorkspaceUuid: workspace.Uuid,
-		Name:          "test",
-		Brief:         "test get brief",
-		Requirements:  "Test get requirements",
-		Architecture:  "Test get architecture",
-		Url:           "Test get url",
-		Priority:      1,
-		Created:       &now,
-		Updated:       &now,
-		CreatedBy:     "test",
-		UpdatedBy:     "test",
-	}
-
-	// create WorkspaceFeatures
-	TestDB.CreateOrEditFeature(workspaceFeatures)
-
-	featurePhase := FeaturePhase{
-		Uuid:        uuid.New().String(),
-		FeatureUuid: workspaceFeatures.Uuid,
-		Name:        "test get feature phase",
-		Priority:    1,
-		Created:     &now,
-		Updated:     &now,
-	}
-
-	// create FeaturePhase
-	TestDB.CreateOrEditFeaturePhase(featurePhase)
-
-	ticket := Tickets{
-		UUID:        uuid.New(),
-		FeatureUUID: workspaceFeatures.Uuid,
-		PhaseUUID:   featurePhase.Uuid,
-		Name:        "test get tickets 1",
-		CreatedAt:   now,
-		UpdatedAt:   now,
-	}
-
-	// create ticket
-	TestDB.CreateOrEditTicket(&ticket)
-
-	ticket2 := Tickets{
-		UUID:        uuid.New(),
-		FeatureUUID: workspaceFeatures.Uuid,
-		PhaseUUID:   featurePhase.Uuid,
-		Name:        "test get tickets 2",
-		CreatedAt:   now,
-		UpdatedAt:   now,
-	}
-
-	// create ticket
-	TestDB.CreateOrEditTicket(&ticket2)
-
-	// should return an empty slice if no tickets are found for the phase
-	t.Run("should return an empty slice if no tickets are found for the phase", func(t *testing.T) {
-		result, err := TestDB.GetTicketsByPhase(uuid.New().String())
-		if err != nil {
-			t.Errorf("expected no error but got %v", err)
-		}
-
-		if len(result) != 0 {
-			t.Errorf("expected an empty slice but got %v", result)
-		}
-	})
-
-	// should return a slice of tickets if the phase exists
-	t.Run("should return a slice of tickets if the phase exists", func(t *testing.T) {
-		result, err := TestDB.GetTicketsByPhase(featurePhase.Uuid)
-		if err != nil {
-			t.Errorf("expected no error but got %v", err)
-		}
-
-		if len(result) != 2 {
-			t.Errorf("expected 2 tickets but got %v", len(result))
-		}
-	})
-}
